ci: Jenkins Setup
refactor: Moved all environment variables to the root of the project
feat: Created Jenkins file with complete pipeline
refactor: Separated postgres volumes into separate docker-compose file
feat: Implemented health check for postgres service due to error encountered when testing

When the postgres service is first initialised, migrations must be applied, however on consecutive runs it causes errors. As such I decided to separate the volumes into a separate file which are used in combination with the main file in development/production.

Jenkinsfile: I created a dedicated agent for this application which contains all the necessary environment variables.
The '-v' when closing the docker services is to remove the unnamed volumes created automatically by the postgres service.