feat: Setup Cypress feat: Added e2e test for Auth0 login feat: Updated .env.example with new variables feat: Added env variables to docker-compose test and dev files feat: Updated Jenkinsfile with new cypress test feat: Replaced docker image in test and dev Docker files for cypress feat: Passed env variables as argument to vite production build feat: Added id of 'login' to LoginButton to target in cypress tests refactor: Switched ports 5173 to 3000 on client side refactor: Switched ports 3000 to 8080 on server side refactor: Moved .gitignore list in client to root
Using .mjs instead of .ts for the cypress config due to error message 'ReferenceError: exports is not defined in ES module scope'.

Vites production build only accepts environment variables set before the build and so I had to pass them as arguments.

Swapped ports on client and server side to conform with standards.