feat(client): Removed AuthContext in favour of event driven updates with local storage feat(client): Cleaned up services a little. Made them default exports. Removed axiosInstance dependency and used axiosInstance directly from utils. feat(client/test): Updated tests using auth services and/or auth provider feat(client): Created hook useAccessToken which listens to local storage for changes feat(client): Updated useAuthCheck to use useAccessToken feat(client): Created utility (and test) for decoding access token feat(client): Created utilities for updating and removing accessToken from local storage with even dispatches feat(client): Updated axios-config.ts to return instantiated axiosInstance feat(server): Updated deleteUser controller to delete all friendRequests containing users ID before deleting user
There were issues I faced with using AuthContext as the source of the users access token. One such issue was that I couldn't refresh the access token unless from inside the context, which meant anytime I used a service that wasn't in there it wouldn't be possible to update the access token. The solution was to move the access token to local storage at the expense of security. Where necessary, changes to the access token in local storage are detected through the use of events.